Database Rework
Reworked how the database is being created.

You now can provide the constructor of a model that's based on the BaseModel class with an ID. If an ID is provided, it will check if a database array exists for that model and create on if not. The model will then store itself inside the database.

Added functionality to indicate on how to solve and dissolve a model. See accounts for an easy example.

Added functionality to save a profile character.