Working Code For Loot Generation in Pockets!
IDs in ItemEntries need to be updated to smaller Items. All of these item IDs are IDs of bigger items that cannot be fit into the pocket. Otherwise, there are no remaining/left errors in the code, and i believe that after changing the IDs, the code will do its job.